WIP: Moved table and collection into own dirs (Will do with other core types as well)
