+++ src/collection/index.ts
@@ -1,0 +1,2 @@
+export * from './collection';
+export * from './collection-constructor';
