+++ src/table/table.ts
@@ -1,22 +1,22 @@
-import { IDBObjectStore, IDBValidKey } from './public/types/indexeddb';
-import { ModifyError, BulkError, errnames, exceptions, fullNameExceptions, mapError } from './errors';
-import { Table as ITable } from './public/types/table';
-import { TableSchema } from './public/types/table-schema';
-import { TableHooks } from './public/types/table-hooks';
-import { DexiePromise as Promise, PSD, newScope, wrap, rejection } from './helpers/promise';
-import Events from './helpers/Events';
-import { hookCreatingChain, nop, pureFunctionChain, mirror, hookUpdatingChain, hookDeletingChain } from './functions/chaining-functions';
-import { Transaction } from './transaction';
-import { Dexie } from './dexie';
-import { tempTransaction } from './functions/temp-transaction';
-import { eventRejectHandler, hookedEventRejectHandler, hookedEventSuccessHandler, BulkErrorHandlerCatchAll, eventSuccessHandler } from './functions/event-wrappers';
-import { WhereClause } from './where-clause';
-import { Collection } from './collection';
-import { isArray, keys, getByKeyPath, hasOwn, setByKeyPath, deepClone, tryCatch, arrayToObject } from './functions/utils';
-import { maxString } from './globals/constants';
-import { combine } from './functions/combine';
-import { PromiseExtended } from "./public/types/promise-extended";
-import { bulkDelete } from './functions/bulk-delete';
+import { IDBObjectStore, IDBValidKey } from '../public/types/indexeddb';
+import { ModifyError, BulkError, errnames, exceptions, fullNameExceptions, mapError } from '../errors';
+import { Table as ITable } from '../public/types/table';
+import { TableSchema } from '../public/types/table-schema';
+import { TableHooks } from '../public/types/table-hooks';
+import { DexiePromise as Promise, PSD, newScope, wrap, rejection } from '../helpers/promise';
+import Events from '../helpers/Events';
+import { hookCreatingChain, nop, pureFunctionChain, mirror, hookUpdatingChain, hookDeletingChain } from '../functions/chaining-functions';
+import { Transaction } from '../transaction';
+import { Dexie } from '../dexie';
+import { tempTransaction } from '../functions/temp-transaction';
+import { eventRejectHandler, hookedEventRejectHandler, hookedEventSuccessHandler, BulkErrorHandlerCatchAll, eventSuccessHandler } from '../functions/event-wrappers';
+import { WhereClause } from '../where-clause';
+import { Collection } from '../collection';
+import { isArray, keys, getByKeyPath, hasOwn, setByKeyPath, deepClone, tryCatch, arrayToObject } from '../functions/utils';
+import { maxString } from '../globals/constants';
+import { combine } from '../functions/combine';
+import { PromiseExtended } from "../public/types/promise-extended";
+import { bulkDelete } from '../functions/bulk-delete';
 
 /** class Table
  *
